+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - Added RBX class property type change detection
     - Rename boosting `estimators` param to `epochs`
     - Neural net-based learners can now train for 0 epochs
+    - Rename Labeled `stratify()` to `stratifyByLabel()`
     
 - 1.2.1
     - Refactor stratified methods on Labeled dataset